ТЗ-8. Перенос методов в TaskManagerTest.
